# Optimize parquet row reader