--- a/column_chunk.go
+++ b/column_chunk.go
@@ -191,113 +191,168 @@ func (r *columnChunkReader) writeRowsTo(w pageAndValueWriter, limit int64) (numR
 }
 */
 
-type columnReadRowFunc func(Row, byte, []columnChunkReader) (Row, error)
+type columnReadRowsFunc func([]Row, byte, []columnChunkReader) (int, error)
 
-func columnReadRowFuncOf(node Node, columnIndex int, repetitionDepth byte) (int, columnReadRowFunc) {
-	var read columnReadRowFunc
+func columnReadRowsFuncOf(node Node, columnIndex int, repetitionDepth byte) (int, columnReadRowsFunc) {
+	var read columnReadRowsFunc
 
 	if node.Repeated() {
 		repetitionDepth++
 	}
 
 	if node.Leaf() {
-		columnIndex, read = columnReadRowFuncOfLeaf(columnIndex, repetitionDepth)
+		columnIndex, read = columnReadRowsFuncOfLeaf(columnIndex, repetitionDepth)
 	} else {
-		columnIndex, read = columnReadRowFuncOfGroup(node, columnIndex, repetitionDepth)
+		columnIndex, read = columnReadRowsFuncOfGroup(node, columnIndex, repetitionDepth)
 	}
 
 	if node.Repeated() {
-		read = columnReadRowFuncOfRepeated(read, repetitionDepth)
+		read = columnReadRowsFuncOfRepeated(read, repetitionDepth)
 	}
 
 	return columnIndex, read
 }
 
 //go:noinline
-func columnReadRowFuncOfRepeated(read columnReadRowFunc, repetitionDepth byte) columnReadRowFunc {
-	return func(row Row, repetitionLevel byte, columns []columnChunkReader) (Row, error) {
-		var err error
+func columnReadRowsFuncOfRepeated(read columnReadRowsFunc, repetitionDepth byte) columnReadRowsFunc {
+	return func(rows []Row, repetitionLevel byte, columns []columnChunkReader) (int, error) {
+		for i := range rows {
+			// Repeated columns have variable number of values, we must process
+			// them one row at a time because we cannot predict how many values
+			// need to be consumed in each iteration.
+			row := rows[i : i+1]
 
-		for {
-			n := len(row)
-
-			if row, err = read(row, repetitionLevel, columns); err != nil {
-				return row, err
+			// The first pass looks for values marking the beginning of a row by
+			// having a repetition level equal to the current level.
+			n, err := read(row, repetitionLevel, columns)
+			if err != nil {
+				// The error here may likely be io.EOF, the read function may
+				// also have successfully read a row, which is indicated by a
+				// non-zero count. In this case, we increment the index to
+				// indicate to the caller than rows up to i+1 have been read.
+				if n > 0 {
+					i++
+				}
+				return i, err
 			}
-			if n == len(row) {
-				return row, nil
+
+			// The read function may return no errors and also read no rows in
+			// case where it had more values to read but none corresponded to
+			// the current repetition level. This is an indication that we will
+			// not be able to read more rows at this stage, we must return to
+			// the caller to let it set the repetition level to its current
+			// depth, which may allow us to read more values when called again.
+			if n == 0 {
+				return i, nil
 			}
 
-			repetitionLevel = repetitionDepth
+			// When we reach this stage, we have successfully read the first
+			// values of a row of repeated columns. We continue consuming more
+			// repeated values until we get the indication that we consumed
+			// them all (the read function returns zero and no errors).
+			for {
+				n, err := read(row, repetitionDepth, columns)
+				if err != nil {
+					return i + 1, err
+				}
+				if n == 0 {
+					break
+				}
+			}
 		}
+		return len(rows), nil
 	}
 }
 
 //go:noinline
-func columnReadRowFuncOfGroup(node Node, columnIndex int, repetitionDepth byte) (int, columnReadRowFunc) {
+func columnReadRowsFuncOfGroup(node Node, columnIndex int, repetitionDepth byte) (int, columnReadRowsFunc) {
 	fields := node.Fields()
+
+	if len(fields) == 0 {
+		return columnIndex, func(_ []Row, _ byte, _ []columnChunkReader) (int, error) {
+			return 0, io.EOF
+		}
+	}
+
 	if len(fields) == 1 {
 		// Small optimization for a somewhat common case of groups with a single
 		// column (like nested list elements for example); there is no need to
 		// loop over the group of a single element, we can simply skip to calling
 		// the inner read function.
-		return columnReadRowFuncOf(fields[0], columnIndex, repetitionDepth)
+		return columnReadRowsFuncOf(fields[0], columnIndex, repetitionDepth)
 	}
 
-	group := make([]columnReadRowFunc, len(fields))
+	group := make([]columnReadRowsFunc, len(fields))
 	for i := range group {
-		columnIndex, group[i] = columnReadRowFuncOf(fields[i], columnIndex, repetitionDepth)
+		columnIndex, group[i] = columnReadRowsFuncOf(fields[i], columnIndex, repetitionDepth)
 	}
 
-	return columnIndex, func(row Row, repetitionLevel byte, columns []columnChunkReader) (Row, error) {
-		var err error
+	return columnIndex, func(rows []Row, repetitionLevel byte, columns []columnChunkReader) (int, error) {
+		// When reading a group, we use the first column as an indicator of how
+		// may rows can be read during this call.
+		n, err := group[0](rows, repetitionLevel, columns)
 
-		for _, read := range group {
-			if row, err = read(row, repetitionLevel, columns); err != nil {
-				break
+		if n > 0 {
+			// Read values for all rows that the group is able to consume.
+			// Getting io.EOF from calling the read functions indicate that
+			// we consumed all values of that particular column, but there may
+			// be more to read in other columns, therefore we must always read
+			// all columns and cannot stop on the first error.
+			for _, read := range group[1:] {
+				_, err2 := read(rows[:n], repetitionLevel, columns)
+				if err2 != nil && !errors.Is(err2, io.EOF) {
+					return 0, err2
+				}
 			}
 		}
 
-		return row, err
+		return n, err
 	}
 }
 
 //go:noinline
-func columnReadRowFuncOfLeaf(columnIndex int, repetitionDepth byte) (int, columnReadRowFunc) {
-	var read columnReadRowFunc
+func columnReadRowsFuncOfLeaf(columnIndex int, repetitionDepth byte) (int, columnReadRowsFunc) {
+	var read columnReadRowsFunc
 
 	if repetitionDepth == 0 {
-		read = func(row Row, _ byte, columns []columnChunkReader) (Row, error) {
+		read = func(rows []Row, _ byte, columns []columnChunkReader) (int, error) {
+			// When the repetition depth is zero, we know that there is exactly
+			// one value per row for this column, and therefore we can consume
+			// as many values as there are rows to fill.
 			col := &columns[columnIndex]
 
-			for {
+			for n := 0; n < len(rows); {
 				if col.offset < len(col.buffer) {
-					row = append(row, col.buffer[col.offset])
+					rows[n] = append(rows[n], col.buffer[col.offset])
+					n++
 					col.offset++
-					return row, nil
+					continue
 				}
 				if err := col.readValues(); err != nil {
-					return row, err
+					return n, err
 				}
 			}
+
+			return len(rows), nil
 		}
 	} else {
-		read = func(row Row, repetitionLevel byte, columns []columnChunkReader) (Row, error) {
+		read = func(rows []Row, repetitionLevel byte, columns []columnChunkReader) (int, error) {
+			// When the repetition depth is not zero, we know that we will be
+			// called with a single row as input. We attempt to read at most one
+			// value of a single row and return to the caller.
 			col := &columns[columnIndex]
 
 			for {
 				if col.offset < len(col.buffer) {
-					if col.buffer[col.offset].repetitionLevel == repetitionLevel {
-						row = append(row, col.buffer[col.offset])
-						col.offset++
+					if col.buffer[col.offset].repetitionLevel != repetitionLevel {
+						return 0, nil
 					}
-					return row, nil
+					rows[0] = append(rows[0], col.buffer[col.offset])
+					col.offset++
+					return 1, nil
 				}
 				if err := col.readValues(); err != nil {
-					if repetitionLevel > 0 && err == io.EOF {
-						err = nil
-					}
-					return row, err
+					return 0, err
 				}
 			}
 		}