Follow up to #183, this PR modifies the implementation of the row reading logic to take advantage of the fact that we can now read batches of rows and avoid traversing abstraction layer where possible.

Here's a snapshot comparing the `parquet.Buffer` benchmark added in #183 with this change:
```
name                old time/op    new time/op    delta
BufferReadRows100x    1.59µs ± 0%    0.87µs ± 0%  -45.57%  (p=0.000 n=9+8)

name                old row/s      new row/s      delta
BufferReadRows100x     62.8M ± 0%    115.4M ± 0%  +83.71%  (p=0.000 n=9+8)
```

In this PR, I also made the following changes:
* simplified the internal `columnCheckReader` and `rowGroupRows` types
* documented the algorithms that reconstruct columns into rows, they were a bit difficult to get back into so I imagine it might be useful to future maintenance
* renamed `columnReadRowFunc` to `readRowsFunc` to be less verbose